+++ src/lib/components/ToastContainer.svelte
@@ -5,7 +5,7 @@
 </script>
 
 <div
-  class="toasters fixed bottom-8 left-1/2 z-[60] flex max-w-[clamp(250px,_50dvw,_500px)] -translate-x-1/2 flex-col items-center gap-2"
+  class="fixed bottom-8 left-1/2 z-[60] flex max-w-[clamp(250px,_50dvw,_500px)] -translate-x-1/2 flex-col items-center gap-2"
 >
   {#each $toastsStore as [id, toastData] (id)}
       <svelte:component this={toastData.componentType} {...toastData.props} />